Added pagination and error handling

+++ src/main/java/pl/pioro/shipmentregister/controller/BranchController.java
@@ -1,8 +1,10 @@
 package pl.pioro.shipmentregister.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import pl.pioro.shipmentregister.entity.Branch;
+import pl.pioro.shipmentregister.exception.SourceNotFoundException;
 import pl.pioro.shipmentregister.repository.BranchRepository;
 
 
@@ -23,23 +25,28 @@ public class BranchController {
     }
 
     @PostMapping(consumes = "application/json")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public Branch create(@RequestBody Branch branch){
         return branchRepository.save(branch);
     }
 
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable("id") long id){
+        if(branchRepository.findById(id) == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         branchRepository.deleteById(id);
     }
 
     @GetMapping(path = "/{id}")
     public Branch findById(@PathVariable("id") long id){
-        return branchRepository.findById(id);
+        Branch branch = branchRepository.findById(id);
+        if(branch == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        return branch;
     }
 
     @PutMapping(path = "/{id}", consumes = "application/json")
     public Branch updateBranch(@PathVariable("id") long id, @RequestBody Branch branch) {
         Branch branchUpdated = branchRepository.findById(id);
+        if(branchUpdated == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         branchUpdated.setName(branch.getName());
         branchUpdated.setCity(branch.getCity());
         branchUpdated.setAddress(branch.getAddress());
