+++ src/main/java/pl/pioro/shipmentregister/controller/CountryController.java
@@ -1,8 +1,10 @@
 package pl.pioro.shipmentregister.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import pl.pioro.shipmentregister.entity.Country;
+import pl.pioro.shipmentregister.exception.SourceNotFoundException;
 import pl.pioro.shipmentregister.repository.CountryRepository;
 
 
@@ -23,23 +25,28 @@ public class CountryController {
     }
 
     @PostMapping(consumes = "application/json")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public Country create(@RequestBody Country country){
         return countryRepository.save(country);
     }
 
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable("id") long id){
+        if(countryRepository.findById(id) == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         countryRepository.deleteById(id);
     }
 
     @GetMapping(path = "/{id}")
     public Country findById(@PathVariable("id") long id){
-        return countryRepository.findById(id);
+        Country country = countryRepository.findById(id);
+        if(country == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        return country;
     }
 
     @PutMapping(path = "/{id}", consumes = "application/json")
     public Country updateCountry(@PathVariable("id") long id, @RequestBody Country country) {
         Country countryUpdated = countryRepository.findById(id);
+        if(countryUpdated == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         countryUpdated.setName(country.getName());
         countryUpdated.setCode(country.getCode());
