+++ src/main/java/pl/pioro/shipmentregister/controller/RecipientController.java
@@ -1,8 +1,11 @@
 package pl.pioro.shipmentregister.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import pl.pioro.shipmentregister.entity.Recipient;
+import pl.pioro.shipmentregister.exception.SourceNotFoundException;
 import pl.pioro.shipmentregister.repository.RecipientRepository;
 
 
@@ -18,28 +21,48 @@ public class RecipientController {
     private RecipientRepository recipientRepository;
 
     @GetMapping
-    public Iterable<Recipient> findAll() {
-        return recipientRepository.findAll();
+    public Iterable<Recipient> findAllActive(@RequestParam(value = "page", required = false) String page, @RequestParam(value = "size", required = false) String size) {
+        if(page != null && size != null) {
+            PageRequest pageRequest = PageRequest.of(Integer.parseInt(page), Integer.parseInt(size));
+            return recipientRepository.findAllByIsActiveTrue(pageRequest);
+        } else {
+            return recipientRepository.findAllByIsActiveTrue();
+        }
+
     }
 
     @PostMapping(consumes = "application/json")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public Recipient create(@RequestBody Recipient recipient){
         return recipientRepository.save(recipient);
     }
 
+    @PatchMapping(path = "/{id}")
+    @ResponseStatus(value = HttpStatus.NO_CONTENT)
+    public void recipientDeactivated(@PathVariable long id){
+        Recipient recipient = recipientRepository.findById(id);
+        if(recipient == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        recipient.setActive(false);
+        recipientRepository.save(recipient);
+    }
+
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable("id") long id){
+        if(recipientRepository.findById(id) == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         recipientRepository.deleteById(id);
     }
 
     @GetMapping(path = "/{id}")
     public Recipient findById(@PathVariable("id") long id){
-        return recipientRepository.findById(id);
+        Recipient recipient = recipientRepository.findById(id);
+        if(recipient == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        return recipient;
     }
 
     @PutMapping(path = "/{id}", consumes = "application/json")
     public Recipient updateRecipient(@PathVariable("id") long id, @RequestBody Recipient recipient) {
         Recipient recipientUpdated = recipientRepository.findById(id);
+        if(recipientUpdated == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         recipientUpdated.setName(recipient.getName());
         recipientUpdated.setAddress(recipient.getAddress());
         recipientUpdated.setCity(recipient.getCity());
