+++ src/main/java/pl/pioro/shipmentregister/controller/ShipmentStatusController.java
@@ -1,8 +1,10 @@
 package pl.pioro.shipmentregister.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import pl.pioro.shipmentregister.entity.ShipmentStatus;
+import pl.pioro.shipmentregister.exception.SourceNotFoundException;
 import pl.pioro.shipmentregister.repository.ShipmentStatusRepository;
 
 import javax.transaction.Transactional;
@@ -22,23 +24,28 @@ public class ShipmentStatusController {
     }
 
     @PostMapping(consumes = "application/json")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public ShipmentStatus create(@RequestBody ShipmentStatus shipmentStatus){
         return shipmentStatusRepository.save(shipmentStatus);
     }
 
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable("id") int id){
+        if(shipmentStatusRepository.findById(id) == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         shipmentStatusRepository.deleteById(id);
     }
 
     @GetMapping(path = "/{id}")
     public ShipmentStatus findById(@PathVariable("id") int id){
-        return shipmentStatusRepository.findById(id);
+        ShipmentStatus shipmentStatus = shipmentStatusRepository.findById(id);
+        if(shipmentStatus == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        return shipmentStatus;
     }
 
     @PutMapping(path = "/{id}", consumes = "application/json")
     public ShipmentStatus updateShipmentStatus(@PathVariable("id") int id, @RequestBody ShipmentStatus shipmentStatus) {
         ShipmentStatus shipmentStatusUpdated = shipmentStatusRepository.findById(id);
+        if(shipmentStatusUpdated == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         shipmentStatusUpdated.setName(shipmentStatus.getName());
 
         return this.shipmentStatusRepository.save(shipmentStatusUpdated);
