+++ src/main/java/pl/pioro/shipmentregister/controller/UserController.java
@@ -1,8 +1,12 @@
 package pl.pioro.shipmentregister.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import pl.pioro.shipmentregister.entity.Project;
 import pl.pioro.shipmentregister.entity.User;
+import pl.pioro.shipmentregister.exception.SourceNotFoundException;
 import pl.pioro.shipmentregister.repository.UserRepository;
 
 import javax.transaction.Transactional;
@@ -17,28 +21,49 @@ public class UserController {
     private UserRepository userRepository;
 
     @GetMapping
-    public Iterable<User> findAll() {
-        return userRepository.findAll();
+    public Iterable<User> findAll(@RequestParam(value = "page", required = false) String page, @RequestParam(value = "size", required = false) String size) {
+        if(page != null && size != null){
+            int pageInt = Integer.parseInt(page);
+            int sizeInt = Integer.parseInt(size);
+            PageRequest findWithPage = PageRequest.of(pageInt, sizeInt);
+            return userRepository.findAll(findWithPage);
+        } else {
+            return userRepository.findAll();
+        }
     }
 
     @PostMapping(consumes = "application/json")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public User create(@RequestBody User user){
         return userRepository.save(user);
     }
 
+    @PatchMapping(path = "/{id}")
+    @ResponseStatus(value = HttpStatus.NO_CONTENT)
+    public void projectDeactivated(@PathVariable long id){
+        User user = userRepository.findById(id);
+        if(user == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        user.setActive(false);
+        userRepository.save(user);
+    }
+
     @DeleteMapping(path = "/{id}")
     public void delete(@PathVariable("id") long id){
+        if(userRepository.findById(id) == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         userRepository.deleteById(id);
     }
 
     @GetMapping(path = "/{id}")
     public User findById(@PathVariable("id") long id){
-        return userRepository.findById(id);
+        User user = userRepository.findById(id);
+        if(user == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
+        return user;
     }
 
     @PutMapping(path = "/{id}", consumes = "application/json")
     public User updateUser(@PathVariable("id") long id, @RequestBody User user) {
         User userUpdated = userRepository.findById(id);
+        if(userUpdated == null) throw new SourceNotFoundException("Source do not found: id= "+ id);
         userUpdated.setName(user.getName());
         userUpdated.setRole(user.getRole());
         userUpdated.setPassword(user.getPassword());
